requirements.txt
numpy
Flask
pillow
onnxruntime
rembg
opencv-python